added move_over_board_recording_potential_wins
This function takes a number of parameters to define its movement,
then moves over the board in the designated way. It then moves this
way. While moving, it checks for playable positions which could,
potentially, win the game for a player on the that players next turn
(if they're not blocked by the other player first). Such positions,
and they player who could potentially win at them, are recorded
on a given n_in_a_row_position_matrix.