removed full_n_in_a_row board checking functions
these functions are not going to be applicable in most cases. They
expect that the n_in_a_row_length will be the same as the
length. While this is true in some cases (3x3 boards, 3 in a row),
it's not true for the generalized case. I'm removing them.
Eventually I should have a generic n_in_a_row checking function